+++ app/src/main/java/uz/androbeck/virtualbank/data/api/AuthenticationService.kt
@@ -3,7 +3,9 @@ package uz.androbeck.virtualbank.data.api
 import retrofit2.http.Body
 import retrofit2.http.POST
 import uz.androbeck.virtualbank.data.dto.common.response.TokenResDto
+import uz.androbeck.virtualbank.data.dto.common.response.sing_up_verify_response.SuccessResDto
 import uz.androbeck.virtualbank.data.dto.request.SignUpReqDto
+import uz.androbeck.virtualbank.data.dto.request.SignUpVerifyReqDto
 import uz.androbeck.virtualbank.utils.Constants
 
 interface AuthenticationService {
@@ -11,4 +13,8 @@ interface AuthenticationService {
     suspend fun signUp(
         @Body request: SignUpReqDto
     ): TokenResDto
+    @POST(Constants.Endpoint.SIGN_UP_VERIFY)
+    suspend fun signUpVerify(
+        @Body request: SignUpVerifyReqDto
+    ): SuccessResDto
 }
