+++ app/src/main/java/uz/androbeck/virtualbank/data/repository/AuthenticationRepository.kt
@@ -2,8 +2,12 @@ package uz.androbeck.virtualbank.data.repository
 
 import kotlinx.coroutines.flow.Flow
 import uz.androbeck.virtualbank.data.dto.common.response.TokenResDto
+import uz.androbeck.virtualbank.data.dto.common.response.sing_up_verify_response.SuccessResDto
 import uz.androbeck.virtualbank.data.dto.request.SignUpReqDto
+import uz.androbeck.virtualbank.data.dto.request.SignUpVerifyReqDto
 
 interface AuthenticationRepository {
     fun signUp(request: SignUpReqDto): Flow<TokenResDto>
+    fun signUpVerify(request: SignUpVerifyReqDto): Flow<SuccessResDto>
+
 }
