+++ app/src/main/java/uz/androbeck/virtualbank/data/source/remote/AuthenticationRemoteDataSource.kt
@@ -2,10 +2,13 @@ package uz.androbeck.virtualbank.data.source.remote
 
 import kotlinx.coroutines.flow.Flow
 import uz.androbeck.virtualbank.data.dto.common.response.TokenResDto
+import uz.androbeck.virtualbank.data.dto.common.response.sing_up_verify_response.SuccessResDto
 import uz.androbeck.virtualbank.data.dto.request.SignUpReqDto
+import uz.androbeck.virtualbank.data.dto.request.SignUpVerifyReqDto
 
 interface AuthenticationRemoteDataSource {
     fun signUp(request: SignUpReqDto): Flow<TokenResDto>
+    fun signUpVerify(request: SignUpVerifyReqDto): Flow<SuccessResDto>
 }
