+++ app/src/main/java/uz/androbeck/virtualbank/data/source/remote/AuthenticationRemoteDataSourceImpl.kt
@@ -3,6 +3,7 @@ package uz.androbeck.virtualbank.data.source.remote
 import kotlinx.coroutines.flow.flow
 import uz.androbeck.virtualbank.data.api.AuthenticationService
 import uz.androbeck.virtualbank.data.dto.request.SignUpReqDto
+import uz.androbeck.virtualbank.data.dto.request.SignUpVerifyReqDto
 import javax.inject.Inject
 
 class AuthenticationRemoteDataSourceImpl @Inject constructor(
@@ -10,5 +11,9 @@ class AuthenticationRemoteDataSourceImpl @Inject constructor(
 ) : AuthenticationRemoteDataSource {
     override fun signUp(request: SignUpReqDto) = flow {
         emit(service.signUp(request))
+    }
+
+    override fun signUpVerify(request: SignUpVerifyReqDto)= flow {
+        emit(service.signUpVerify(request))
     }
 }
