+++ app/src/main/java/uz/androbeck/virtualbank/di/NetworkModule.kt
@@ -15,7 +15,6 @@ import retrofit2.Converter
 import retrofit2.Retrofit
 import uz.androbeck.virtualbank.BuildConfig
 import uz.androbeck.virtualbank.data.api.AuthenticationService
-import uz.androbeck.virtualbank.data.api.SignUpVerifyService
 import uz.androbeck.virtualbank.network.ErrorHandlingCallAdapterFactory
 import uz.androbeck.virtualbank.network.errors.ErrorHandler
 import uz.androbeck.virtualbank.network.errors.ErrorHandlerImpl
@@ -69,7 +68,7 @@ object NetworkModule {
     fun provideRetrofit(
         okHttpClient: OkHttpClient,
         converter: Converter.Factory,
-        callFactory: ErrorHandlingCallAdapterFactory
+        callFactory: ErrorHandlingCallAdapterFactory,
     ): Retrofit {
         return Retrofit.Builder()
             .baseUrl(BuildConfig.BASE_URL)
@@ -82,15 +81,8 @@ object NetworkModule {
     @Provides
     @Singleton
     fun provideAuthService(
-        retrofit: Retrofit
+        retrofit: Retrofit,
     ): AuthenticationService = retrofit.create(AuthenticationService::class.java)
-
-    @Provides
-    @Singleton
-    fun provideSignUpVerifyService(
-        retrofit: Retrofit
-    ): SignUpVerifyService = retrofit.create(SignUpVerifyService::class.java)
-
 
     @Provides
     fun provideErrorHandler(errorHandlerImpl: ErrorHandlerImpl): ErrorHandler {
