+++ app/src/main/java/uz/androbeck/virtualbank/di/UseCaseModule.kt
@@ -5,13 +5,9 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
 import uz.androbeck.virtualbank.data.repository.AuthenticationRepository
-import uz.androbeck.virtualbank.data.repository.sign_up_verify_repository.SignUpVerifyRepository
 import uz.androbeck.virtualbank.domain.mapper.auth.SignUpMapper
 import uz.androbeck.virtualbank.domain.mapper.auth.TokenMapper
-import uz.androbeck.virtualbank.domain.mapper.auth.segn_up_verify.SignUpVerifyMapper
-import uz.androbeck.virtualbank.domain.mapper.auth.segn_up_verify.SignUpVerifyTokenMapper
 import uz.androbeck.virtualbank.domain.useCase.authentication.SignUpUseCase
-import uz.androbeck.virtualbank.domain.useCase.authentication.SignUpVerifyUseCase
 
 @Module
 @InstallIn(ViewModelComponent::class)
@@ -23,17 +19,4 @@ object UseCaseModule {
         tokenMapper: TokenMapper,
         signUpMapper: SignUpMapper,
     ) = SignUpUseCase(authenticationRepository, tokenMapper, signUpMapper)
-
-    @Provides
-    fun providesSignUpVerifyUseCase(
-        signUpVerifyRepository: SignUpVerifyRepository,
-        signUpVerifyTokenMapper: SignUpVerifyTokenMapper,
-        signUpVerifyMapper: SignUpVerifyMapper,
-    ): SignUpVerifyUseCase {
-        return SignUpVerifyUseCase(
-            signUpVerifyRepository = signUpVerifyRepository,
-            tokenMapper = signUpVerifyTokenMapper,
-            signUpVerifyMapper = signUpVerifyMapper
-        )
-    }
 }
