+++ app/src/main/java/uz/androbeck/virtualbank/domain/useCase/authentication/SignUpVerifyUseCase.kt
@@ -2,7 +2,7 @@ package uz.androbeck.virtualbank.domain.useCase.authentication
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import uz.androbeck.virtualbank.data.repository.sign_up_verify_repository.SignUpVerifyRepository
+import uz.androbeck.virtualbank.data.repository.AuthenticationRepository
 import uz.androbeck.virtualbank.domain.mapper.auth.segn_up_verify.SignUpVerifyMapper
 import uz.androbeck.virtualbank.domain.mapper.auth.segn_up_verify.SignUpVerifyTokenMapper
 import uz.androbeck.virtualbank.domain.ui_models.authentication.SignUpVerifyReqUIModel
@@ -10,12 +10,12 @@ import uz.androbeck.virtualbank.domain.ui_models.authentication.SignUpVerifyToke
 import javax.inject.Inject
 
 class SignUpVerifyUseCase @Inject constructor(
-    private val signUpVerifyRepository: SignUpVerifyRepository,
+    private val authenticationRepository: AuthenticationRepository,
     private val tokenMapper: SignUpVerifyTokenMapper,
     private val signUpVerifyMapper: SignUpVerifyMapper,
 ) {
     operator fun invoke(uiReqModel: SignUpVerifyReqUIModel): Flow<SignUpVerifyTokenUIModel> {
         val request = signUpVerifyMapper.toDTO(uiReqModel)
-        return signUpVerifyRepository.signUpVerify(request).map { tokenMapper.toUIModel(it) }
+        return authenticationRepository.signUpVerify(request).map { tokenMapper.toUIModel(it) }
     }
 }
